+++ README.md
@@ -1,5 +1,8 @@
 # 자동매매 봇 만들기
-업비트와 키움증권(혹은 크레온) API를 이용한 자동 매매봇 만들기
+업비트와 ~~키움증권(혹은 크레온)~~ API를 이용한 자동 매매봇 만들기
+
+- [ ] 각 전략별 시드 나누기 어떻게 할까
+
 
 # 매매전략
 ### 1. 무한매수법
@@ -7,9 +10,29 @@
 [이 글](https://m.blog.naver.com/edgar0418/222224056120)을 참조하여 만들 것.  
 우선은 간단한 매매방식부터 만들어서 적용시켜 볼 것이다
 - [x] ~~API발급 받기~~
-- [ ] 해외 주식 거래가 가능한 API인지 문의
-- [x] 업비트 정액 매매 가능한지 문의
-- [x] 파이썬 스케쥴러 활용하여 같은 시간에 매수, 매도 하도록
-- [ ] 호스팅 어떻게?
-- [ ] 매매가 이루어지면 알림 가능? --> Telegram
-- [ ] 24시간 / 더 큰 변동성을 가진 코인에 맞게 수정 필요
+- [x] ~~해외 주식 거래가 가능한 API인지 문의~~
+- [ ] 업비트 정액 매매 가능한지 문의
+- [x] ~~파이썬 스케쥴러 활용하여 같은 시간에 매수, 매도 하도록~~
+- [x] ~~호스팅 어떻게?~~
+- [x] ~~매매가 이루어지면 알림 가능? --> Telegram~~
+- [x] ~~24시간 / 더 큰 변동성을 가진 코인에 맞게 수정 필요~~
+- [ ] 텔레그램 응답형 추가 / 코드수정 필요
+
+### 2. 변동성 돌파전략
+래리 윌러엄스의 변동성 돌파전략  
+
+1. 레인지 계산: 기준시각으로부터 전일 고거에서 전일 저가를 뺀다. (하루 안에 움직인 가격의 최대폭)
+1. 매수 기준: 시가 기준으로 가격이 'range * k' 이상 상승하면 해당 가격에 매수
+    - k는 매수 횟수와 관련됨 --> 0.5 추천
+    - k가 높으면 돌파 기준이 높아지므로 크게 상승할때만 매수하게 됨
+1. 매도 기준: 당일 종가에 매도
+1. 종가, 고가, 저가 기준
+    - 시가: 아무때나 적절한 시간 선택
+    - 종가: 시가로부터 24시간 후
+    - 고가(저가): 정한 시가 기준 24시간 동안 가장 높은(낮은) 가격
+
+
+### 3. 백테스트
+다양한 전략들을 테스트하기 위한 툴 
+- [ ] ⭐️⭐️⭐️ 과거 코인 가격 데이터 구하기 ⭐️⭐️⭐️ 
+
